doc.markdown: First makrdown language draft
The mardown language is intended to be used as generation target
for other languages.

In the mid-long term it will replace the existing documentation
generators.

This implementation covers the basic syntax as described in:
https://www.markdownguide.org/cheat-sheet/

doc.gen_markdown: documentation to markdown generator

Implemented transformation from documentation language to
markdown.

At the moment only the documentation elements that have a
correspondence in markdown basic syntax are generated.

A make facet processes the generated markdown and produces
an html file.